% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorConv.R
\name{ColorConv}
\alias{ColorConv}
\title{Image Color Conversion}
\usage{
ColorConv(img, color)
}
\arguments{
\item{img}{path of the input image from package jpeg's function readJPEG().}

\item{color}{desired color to convert: gray, red, green, blue}
}
\value{
The converted image
}
\description{
This function converts an image to one of the color : gray, red, green, blue
}
\examples{
img <- array(runif(1000),dim=c(10,10,3))
ColorConv(img, color = 'gray')

}
