% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImgRotate.r
\name{ImgRotate}
\alias{ImgRotate}
\title{Rotate an image 90, 180, 270, or 360 degrees counterclockwise.}
\usage{
ImgRotate(img, degree)
}
\arguments{
\item{img, }{path of the input image}

\item{degree, }{desired degree to rotate img: 90, 180, 270, 360.}
}
\value{
rotated_img
}
\description{
Rotate an image 90, 180, 270, or 360 degrees counterclockwise.
}
\examples{
img <- array(runif(1000),dim=c(10,10,3))
ImgRotate(img, 270)
}
