% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crop.R
\name{Crop}
\alias{Crop}
\title{Title Crop}
\usage{
Crop(input_path, output_path, height, width)
}
\arguments{
\item{input_path}{----String, path of the input image}

\item{output_path}{----String, path of the output image}

\item{height}{----Int, height of the desired image}

\item{width}{----Int, width of the desired image}
}
\description{
Title Crop
}
